GPS: Changes to add a delay in XTRA data injection, increase block size
Adds a delay of 10 ms in the loop used to inject XTRA data so that the
load on RPC task on the modem is reduced. Otherwise there is a chance
that RPC task on the modem may starve the modem processor. The block
size has to be increase to minimize the delay introduced to TTFF when
XTRA is used.

CRs-Fixed:292573

Change-Id: I746ab0072f20802354a66499f7abbf1396c24a51