GPS: Enhanced the HAL layer for V2 Non-trusted call flow
In V2 Non-trusted call flow the gps engine will request two data
connections (1- MPC & 1- PDE) .The HAL layer is redesigned to manage
these internally so that we send one data connection open request to
the location framework when we get the first data open request
from the engine. All subsequent data open requests are ignored and
in the same principle the HAL layer will accumulate all data connection
close requests so that a single request can be sent to the framework when
the last close request from the engine is received.

Change-Id: I401422d84e653a6b074b3d3719c0a1e7fa57940b
CRs-Fixed: 231266